% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha}
\alias{alpha}
\title{Alpha calculation for each raster cell}
\usage{
alpha(bin, type = NULL, cores = 1, filename = NULL, ...)
}
\arguments{
\item{bin}{A SpatRaster with presence-absence data (0 or 1) for a set of species.}

\item{type}{It can be a phylo object with phylogenetic tree for a set of species or a data.frame object with species traits.}

\item{cores}{A positive integer. If cores > 1, a 'parallel' package cluster with that many cores is created and used.}

\item{filename}{A character. Output filename.}

\item{...}{Additional arguments to be passed passed down from a calling function.}
}
\value{
A SpatRaster with alpha result.
}
\description{
Alpha calculation for each raster cell
}
\examples{
\dontrun{
set.seed(100)
bin <- terra::rast(array(sample(c(rep(1, 800), rep(0, 200))), dim = c(10, 10, 10)))
names(bin) <- paste0("sp", 1:10)
set.seed(100)
mass <- runif(10, 10, 800)
beak.size <- runif(10, .2, 5)
tail.length <- runif(10, 2, 10)
wing.length <- runif(10, 15, 60)
range.size <- runif(10, 10000, 100000)
traits <- data.frame(mass, beak.size, tail.length, wing.length, range.size)
rownames(traits) <- names(bin)
set.seed(100)
tree <- ape::rtree(n = 10, tip.label = paste0("sp", 1:10))
alpha.td <- alpha(bin)
alpha.td
alpha.fd <- alpha(bin, traits)
alpha.fd
alpha.pd <- alpha(bin, tree)
alpha.pd
}
}
