% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphadv_fd.R
\name{alpha_fd}
\alias{alpha_fd}
\title{Functional alpha diversity calculation}
\usage{
alpha_fd(r, traits, filename = NULL, ...)
}
\arguments{
\item{r}{SpatRaster object with binarized distribution projected to all species for a given climate scenario}

\item{traits}{data.frame object with traits as columns and species as rownames}

\item{filename}{Output filename}

\item{...}{Additional arguments to be passed passed down from a calling function}
}
\value{
SpatRaster object with functional diversity calculates as functional richness
}
\description{
It calculates functional richness for a given climate scenario
}
\examples{
\dontrun{
set.seed(100)
ref <- terra::rast(array(sample(c(rep(1, 800), rep(0, 200))), dim = c(20, 20, 10)))
names(ref) <- paste0("sp", 1:10)
ref
set.seed(100)
mass <- runif(10, 10, 800)
beak.size <- runif(10, .2, 5)
tail.length <- runif(10, 2, 10)
wing.length <- runif(10, 15, 60)
range.size <- runif(10, 10000, 100000)
traits <- as.data.frame(cbind(mass, beak.size, tail.length, wing.length, range.size))
rownames(traits) <- paste0("sp", 1:10)
traits
alpha.fd <- alpha_fd(ref, traits)
alpha.fd
}
}
