% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphadv.R
\name{alphadv}
\alias{alphadv}
\title{Alpha diversity calculation for raster}
\usage{
alphadv(bin, filename = NULL, cores = 1, ...)
}
\arguments{
\item{bin}{SpatRaster object with binarized distribution projected to all species for a given climate scenario}

\item{filename}{Output filename}

\item{cores}{A positive integer indicating if parallel processing should be used (cores > 1)}

\item{...}{Additional arguments to be passed down from a calling function}
}
\value{
SpatRaster object with alpha diversity/species richness
}
\description{
Alpha diversity calculation for raster
}
\examples{
\dontrun{
set.seed(100)
ref <- terra::rast(array(sample(c(rep(1, 800), rep(0, 200))), dim = c(10, 10, 10)))
names(ref) <- paste0("sp", 1:10)
ref
rich <- alphadv(ref)
rich
}
}
