% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_spatial.R
\name{beta.focal.spat}
\alias{beta.focal.spat}
\title{Compute beta diversity on spatRast objects}
\usage{
beta.focal.spat(
  x,
  fm = NULL,
  d = mean(terra::res(terra::rast(x))) * 2,
  type = "circle",
  index.family = "sorensen",
  tree = NA,
  filetype = "GTiff",
  filename = NULL,
  overwrite = T,
  numCores = 1,
  ...
)
}
\arguments{
\item{x}{raster brick or raster stack of species presence/absence}

\item{fm}{focal matrix. Numeric. Make a focal ("moving window")
weight matrix for use in the focal function}

\item{d}{numeric. If \code{type=circle}, the radius of the circle (in units of the crs). If \code{type=rectangle} the dimension of the rectangle (one or two numbers). If \code{type=Gauss} the size of sigma, and optionally another number to determine the size of the matrix returned (default is 3*sigma)}

\item{type}{character indicating the type of filter to be returned}

\item{index.family}{ family of dissimilarity indices, partial match of \code{"sorensen"} or \code{"jaccard"}.}

\item{tree}{ a phylogenetic tree of class phylo with tips names identic to species names from the community matrix.}

\item{filetype}{file format expresses as GDAL driver names. If this
argument is not supplied, the driver is derived from the filename.
For details see \code{\link[terra]{writeRaster}}}

\item{filename}{character. Output filename. Can be a single filename, or as many filenames as \code{nlyr(x)} to write a file for each layer}

\item{overwrite}{Logical. Should saved files be overwritten with new values?}

\item{numCores}{Number of cores to be used in parallel calculation}

\item{...}{additional arguments for for writing files. See Details }
}
\description{
This function will compute beta diversity on spatRast objects
that contain binary (presence/absence) species distribution data
}
