% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_spatial.R
\name{beta.focal.vec}
\alias{beta.focal.vec}
\title{Compute beta diversity on vectors}
\usage{
beta.focal.vec(x, nspp, index.family = "sorensen", tree = NA)
}
\arguments{
\item{x}{numeric vector with multiple species and multiple sites}

\item{nspp}{numeric. number of species to transform the vector
into a matrix with nspp columns}

\item{index.family}{ family of dissimilarity indices, partial match of \code{"sorensen"} or \code{"jaccard"}.}

\item{tree}{ a phylogenetic tree of class phylo with tips names identic to species names from the community matrix.}
}
\description{
This function will compute beta diversity on vectors containing
multiple species at multiple sites. Species at sites should be
placed sequentially, so that the vector can be transformed in a
matrix with species at columns and sites at rows
}
