% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_spatial.R
\name{beta.temp.spat}
\alias{beta.temp.spat}
\title{Compute beta diversity on spatRast objects}
\usage{
beta.temp.spat(
  x,
  y,
  index.family = "sorensen",
  filetype = "GTiff",
  filename = NULL,
  overwrite = T,
  numCores = 1,
  ...
)
}
\arguments{
\item{x}{raster brick or raster stack of species presence/absence}

\item{y}{raster brick or raster stack of species presence/absence}

\item{index.family}{ family of dissimilarity indices, partial match of \code{"sorensen"} or \code{"jaccard"}.}

\item{filetype}{file format expresses as GDAL driver names. If this
argument is not supplied, the driver is derived from the filename.
For details see \code{\link[terra]{writeRaster}}}

\item{filename}{character. Output filename}

\item{overwrite}{Logical. Should saved files be overwritten with new values?}

\item{numCores}{Number of cores to be used in parallel calculation}

\item{...}{additional arguments for \code{fun}. These are typically numerical constants. They should *never* be another SpatRaster}
}
\description{
This function will compute beta diversity on spatRast objects
that contain binary (presence/absence) species distribution data
}
