% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betatemp_fd.R
\name{betatemp_fd}
\alias{betatemp_fd}
\title{Beta temporal of functional diversity}
\usage{
betatemp_fd(ref, fut, traits, cores = 1, filename = NULL, ...)
}
\arguments{
\item{ref}{A SpatRaster object with binary maps of species distributions}

\item{fut}{A SpatRaster object with binary maps of species distributions}

\item{traits}{A data.frame object with traits as columns and species as rownames}

\item{cores}{A positive integer indicating if parallel processing should be used (cores > 1)}

\item{filename}{Output filename}

\item{...}{Additional arguments to be passed down from a calling function}
}
\value{
A SpatRaster object with beta total, turnover, and nestedness
}
\description{
Beta temporal of functional diversity
}
\examples{
\dontrun{
set.seed(100)
ref <- terra::rast(array(sample(c(rep(1, 800), rep(0, 200))), dim = c(10, 10, 10)))
names(ref) <- paste0("sp", 1:10)
fut <- terra::rast(array(sample(c(rep(1, 400), rep(0, 600))), dim = c(10, 10, 10)))
names(fut) <- names(ref)
set.seed(100)
mass <- runif(10, 10, 800)
beak.size <- runif(10, .2, 5)
tail.length <- runif(10, 2, 10)
wing.length <- runif(10, 15, 60)
range.size <- runif(10, 10000, 100000)
traits <- data.frame(mass, beak.size, tail.length, wing.length, range.size)
rownames(traits) <- names(ref)
beta.fd <- betatemp_fd(ref, fut, traits)
beta.fd
}
}
