% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betatemp_td.R
\name{betatemp_td}
\alias{betatemp_td}
\title{Beta temporal of taxonomic diversity}
\usage{
betatemp_td(ref, fut, cores = 1, filename = NULL, ...)
}
\arguments{
\item{ref}{A SpatRaster object with binary maps of species distributions}

\item{fut}{A SpatRaster object with binary maps of species distributions}

\item{cores}{A positive integer indicating if parallel processing should be used (cores > 1)}

\item{filename}{Output filename}

\item{...}{Additional arguments to be passed down from a calling function}
}
\value{
A SpatRaster object with beta total, turnover, and nestedness
}
\description{
Beta temporal of taxonomic diversity
}
\examples{
\dontrun{
set.seed(100)
ref <- terra::rast(array(sample(c(rep(1, 800), rep(0, 200))), dim = c(10, 10, 10)))
names(ref) <- paste0("sp", 1:10)
fut <- terra::rast(array(sample(c(rep(1, 400), rep(0, 600))), dim = c(10, 10, 10)))
names(fut) <- names(ref)
beta.td <- betatemp_td(ref, fut)
beta.td
}
}
