% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betatempdv.R
\name{betatempdv}
\alias{betatempdv}
\title{Temporal beta diversity calculation}
\usage{
betatempdv(ref, fut, filename = NULL, cores = 1, ...)
}
\arguments{
\item{ref}{SpatRaster object with binarized distribution projected to all species from climate scenario 1}

\item{fut}{SpatRaster object with binarized distribution projected to all species from climate scenario 2}

\item{filename}{Output filename}

\item{cores}{A positive integer indicating if parallel processing should be used (cores > 1)}

\item{...}{Additional arguments to be passed down from a calling function}
}
\value{
SpatRaster object with each metric as an individual layer (beta total, turnover, nestedness, and ratio)
}
\description{
It calculates beta diversity based on Jaccard dissimilarity index between reference and future scenarios
}
\examples{
\dontrun{
set.seed(100)
ref <- terra::rast(array(sample(c(rep(1, 800), rep(0, 200))), dim = c(10, 10, 10)))
names(ref) <- paste0("sp", 1:10)
ref
fut <- terra::rast(array(sample(c(rep(1, 400), rep(0, 600))), dim = c(10, 10, 10)))
names(fut) <- paste0("sp", 1:10)
fut
b <- betatempdv(ref, fut)
b
}
}
