% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_centroids.R
\name{d.centroids}
\alias{d.centroids}
\title{Function to calculate distance and direction of change between centroids}
\usage{
d.centroids(..., ref = 1)
}
\arguments{
\item{...}{N number of single or multilayer spatrasters}

\item{ref}{Spatraster of reference}
}
\value{
A data frame with distance and direction values
}
\description{
Function to calculate distance and direction of change between centroids
}
\examples{
\donttest{
library(terra)
r1 <- terra::rast(system.file("extdata", "ref.tif",
package = "divraster"))
r2 <- terra::rast(system.file("extdata", "fut.tif",
package = "divraster"))
d.centroids(r1, r2)
}
}
