% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_centroids.R
\name{dd.calc}
\alias{dd.calc}
\title{Function to calculate distance and direction of change between centroids}
\usage{
dd.calc(raster1, raster2)
}
\arguments{
\item{raster1}{A binary spatraster.}

\item{raster2}{A binary spatraster.}
}
\value{
A data frame with distance and direction.
}
\description{
Function to calculate distance and direction of change between centroids
}
\examples{
\donttest{
library(terra)
r1 <- terra::rast(system.file("extdata", "ref.tif",
package = "divraster"))
r2 <- terra::rast(system.file("extdata", "fut.tif",
package = "divraster"))
dd.calc(r1, r2)
}
}
