% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_rast.R
\name{differ.rast}
\alias{differ.rast}
\title{Difference between raster objects}
\usage{
differ.rast(r1, r2, perc = TRUE, filename = "")
}
\arguments{
\item{r1}{A SpatRaster object.}

\item{r2}{A SpatRaster object.}

\item{perc}{Boolean. Default is TRUE to calculate the percentage
of change between r1 and r2. FALSE gives the absolute number
instead.}

\item{filename}{Character. Save results if a name is provided.}
}
\value{
A SpatRaster object with the difference between
r1 and r2.
}
\description{
Difference between raster objects
}
\examples{
\donttest{
library(terra)
rich1 <- terra::rast(system.file("extdata", "rich_ref.tif",
package = "divraster"))
rich2 <- terra::rast(system.file("extdata", "rich_fut.tif",
package = "divraster"))
differ.rast(rich1, rich2)
}
}
