% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_centroids.R
\name{dist.centroids}
\alias{dist.centroids}
\title{Function to calculate distances between centroids of multiple SpatRasters}
\usage{
dist.centroids(..., ref = 1)
}
\arguments{
\item{...}{N number of single or multilayer spatrasters}

\item{ref}{Spatraster of reference}
}
\value{
A data frame with distance values
}
\description{
Function to calculate distances between centroids of multiple SpatRasters
}
\examples{
\donttest{
library(terra)
r1 <- terra::rast(system.file("extdata", "ref.tif",
package = "divraster"))
r2 <- terra::rast(system.file("extdata", "fut.tif",
package = "divraster"))
dist.centroids(r1, r2)
}
}
