% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betatempdv.R
\name{.abc}
\alias{.abc}
\title{Beta diversity parameters}
\usage{
.abc(r, f)
}
\arguments{
\item{r}{Binary vector of reference climate scenario}

\item{f}{Binary vector of future climate scenario}
}
\value{
SpatRaster of beta diversity parameters
}
\description{
Beta diversity parameters
}
\examples{
\dontrun{
set.seed(100)
ref <- terra::rast(array(sample(c(rep(1, 800), rep(0, 200))), dim = c(10, 10, 10)))
names(ref) <- paste0("sp", 1:10)
ref
fut <- terra::rast(array(sample(c(rep(1, 400), rep(0, 600))), dim = c(10, 10, 10)))
names(fut) <- paste0("sp", 1:10)
fut
pars <- .abc(ref, fut)
pars
}
}
