% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{.alpha.vec}
\alias{.alpha.vec}
\title{Alpha calculation for a vector}
\usage{
.alpha.vec(x, type)
}
\arguments{
\item{x}{A SpatRaster with presence-absence data (0 or 1) for a set of species.}

\item{type}{It can be a phylo object with phylogenetic tree for a set of species or a data.frame object with species traits.}
}
\value{
A numeric vector with alpha result.
}
\description{
Alpha calculation for a vector
}
