% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphadv.R
\name{.alphadv}
\alias{.alphadv}
\title{Alpha diversity calculation for vector}
\usage{
.alphadv(bin)
}
\arguments{
\item{bin}{Object of class SpatRaster with binarized distribution projected to all species from climate scenario 1}
}
\value{
Object of class SpatRaster with alpha diversity/species richness
}
\description{
Alpha diversity calculation for vector
}
\examples{
\dontrun{
set.seed(100)
ref <- terra::rast(array(sample(c(rep(1, 750), rep(0, 250))), dim = c(20, 20, 10)))
names(ref) <- paste0("sp", 1:10)
ref
ref.rich <- alphadv(ref)
ref.rich
}
}
