% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_temporal.R
\name{.beta.temporal.vec}
\alias{.beta.temporal.vec}
\title{Beta temporal calculation for a vector}
\usage{
.beta.temporal.vec(x, nspp, spp, type)
}
\arguments{
\item{x}{A numeric vector with presence-absence data (0 or 1) for a set of species.}

\item{nspp}{Number of species.}

\item{spp}{Species names.}

\item{type}{It can be a phylo object with phylogenetic tree for a set of species or a data.frame object with species traits.}
}
\value{
A numeric vector with beta total, turnover, and nestedness.
}
\description{
Beta temporal calculation for a vector
}
