% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadmota.R
\name{loadmota}
\alias{loadmota}
\title{Load data from Mota et al. 2022}
\usage{
loadmota()
}
\value{
List with climate scenarios and suitable area of reference climate scenario by species
}
\description{
Available at \url{https://doi.org/10.1111/ddi.13602}
}
\examples{
\dontrun{
data <- loadmota()
data
}
}
