% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range.list}
\alias{range.list}
\title{Area calculation}
\usage{
\method{range}{list}(x, unit = "km", r2 = NULL, r3 = NULL, filename = "")
}
\arguments{
\item{x}{A list containing multiple SpatRaster with
presence-absence data (0 or 1) for a set of species.}

\item{unit}{Character. Default is "km", but "m" and "ha"
are also available.}

\item{r2}{A SpatRaster with the same extent as "x".}

\item{r3}{A SpatRaster with the same extent as "x".}

\item{filename}{Character. Save results if a name is provided.}
}
\value{
A data.frame with area values in the chosen unit.
}
\description{
Area calculation
}
