% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range.rast}
\alias{range.rast}
\title{Area calculation}
\usage{
\method{range}{rast}(x, unit = "km", ...)
}
\arguments{
\item{x}{A SpatRaster with presence-absence data (0 or 1) for a
given species.}

\item{unit}{Character. Default is "km", "m" and "ha" also
available.}

\item{...}{Additional arguments to be passed passed down from
a calling function.}
}
\value{
A vector with an area in the chosen unit.
}
\description{
Area calculation
}
