% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_fd.R
\name{ses_fd}
\alias{ses_fd}
\title{Standardized Functional Richness calculation for raster}
\usage{
ses_fd(fd, aleats = 999, cores = 1, filename = NULL, ...)
}
\arguments{
\item{fd}{A SpatRaster object with functional richness}

\item{aleats}{A positive integer indicating how many times the calculation should be repeated}

\item{cores}{A positive integer indicating if parallel processing should be used (cores > 1)}

\item{filename}{Output filename}

\item{...}{Additional arguments to be passed down from a calling function}
}
\value{
A SpatRaster object with mean, standard deviation, observed functional richness, and standardized functional richness
}
\description{
Standardized Functional Richness calculation for raster
}
\examples{
\dontrun{
set.seed(100)
ref <- terra::rast(array(sample(c(rep(1, 800), rep(0, 200))), dim = c(10, 10, 10)))
names(ref) <- paste0("sp", 1:10)
ref
set.seed(100)
mass <- runif(10, 10, 800)
beak.size <- runif(10, .2, 5)
tail.length <- runif(10, 2, 10)
wing.length <- runif(10, 15, 60)
range.size <- runif(10, 10000, 100000)
traits <- data.frame(mass, beak.size, tail.length, wing.length, range.size)
rownames(traits) <- paste0("sp", 1:10)
traits
fd <- alpha_fd(ref, traits)
fd
ses.fd <- ses_fd(fd)
ses.fd
}
}
