% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_spatial.R
\name{spat.beta}
\alias{spat.beta}
\title{Spatial beta diversity for raster}
\usage{
spat.beta(
  x,
  tree,
  filename = "",
  global = FALSE,
  fm = NULL,
  d = mean(terra::res(terra::rast(x))) * 2,
  type = "circle",
  na.policy = "omit",
  ...
)
}
\arguments{
\item{x}{A SpatRaster with presence-absence data (0 or 1) for a
set of species.}

\item{tree}{It can be a 'data.frame' with species traits or a
'phylo' with a rooted phylogenetic tree. Species names in 'tree'
and 'x' must match!}

\item{filename}{Character. Save results if a name is provided.}

\item{global}{Logical. Mean of pairwise comparisons between
focal cell and its neighbors (default) or mean of all pairwise
comparisons.}

\item{fm}{Numeric. Focal matrix ("moving window").}

\item{d}{Window radius to compute beta diversity.}

\item{type}{Character. Window format. Default = "circle".}

\item{na.policy}{Character. Default = "omit".}

\item{...}{Additional arguments passed to lower-level functions.}
}
\value{
A SpatRaster with beta results (total, replacement,
richness difference, and ratio).
}
\description{
Calculates spatial beta diversity for
taxonomic (TD), functional (FD), and phylogenetic (PD)
dimensions. Adapted from \code{\link[BAT]{beta}}.
}
