% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{spat.rand}
\alias{spat.rand}
\title{standardize effect size}
\usage{
spat.rand(x, tree, aleats, random = "spat", cores = 1, filename = NULL, ...)
}
\arguments{
\item{x}{SpatRaster}

\item{tree}{species traits or phylogenetic tree}

\item{aleats}{number of randomizations}

\item{random}{randomization method}

\item{cores}{multi-core processing}

\item{filename}{output name}

\item{...}{additional arguments}
}
\value{
SpatRaster with mean, sd, observed, and ses
}
\description{
standardize effect size
}
