% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{spat.trait.vec}
\alias{spat.trait.vec}
\title{Average trait calculation for vector}
\usage{
spat.trait.vec(x, col_trait, ...)
}
\arguments{
\item{x}{A numeric vector with presence-absence data (0 or 1)
for a set of species.}

\item{col_trait}{A numeric vector with trait numbers.}

\item{...}{Additional arguments to be passed passed down from a
calling function.}
}
\value{
Vector of average trait.
}
\description{
Average trait calculation for vector
}
