% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suit_change.R
\name{suit.change}
\alias{suit.change}
\title{Species suitability change between climate scenarios}
\usage{
suit.change(r1, r2, filename = "")
}
\arguments{
\item{r1}{A SpatRaster with presence-absence data (0 or 1)
for a set of species.}

\item{r2}{A SpatRaster with presence-absence data (0 or 1)
for a set of species.}

\item{filename}{Character. Save results if a name is provided.}
}
\value{
A SpatRaster with suitability change (gain, loss,
unchanged, and unsuitable).
}
\description{
Species suitability change between climate scenarios
}
\examples{
#' \donttest{
library(terra)
r1 <- terra::rast(system.file("extdata", "ref.tif",
package = "divraster"))
r2 <- terra::rast(system.file("extdata", "fut.tif",
package = "divraster"))
suit.change(r1, r2)
}
}
