% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_temporal.R
\name{temp.beta}
\alias{temp.beta}
\title{Temporal beta diversity calculation for raster}
\usage{
temp.beta(bin1, bin2, tree, filename = NULL, cores = 1, ...)
}
\arguments{
\item{bin1}{A SpatRaster with presence-absence data (0 or 1) for a set of species.}

\item{bin2}{A SpatRaster with presence-absence data (0 or 1) for a set of species.}

\item{tree}{It can be a data frame with species traits or a phylogenetic tree.}

\item{filename}{Character. Save results if a name is provided.}

\item{cores}{A positive integer. If cores > 1, a 'parallel' package cluster with that many cores is created and used.}

\item{...}{Additional arguments to be passed passed down from a calling function.}
}
\value{
A SpatRaster with beta results.
}
\description{
Temporal beta diversity calculation for raster
}
\examples{
\dontrun{
library(terra)
set.seed(100)
bin1 <- rast(ncol = 5, nrow = 5, nlyr = 10)
values(bin1) <- round(runif(ncell(bin1) * nlyr(bin1)))
names(bin1) <- paste0("sp", 1:10)
bin2 <- rast(ncol = 5, nrow = 5, nlyr = 10)
values(bin2) <- round(runif(ncell(bin2) * nlyr(bin2)))
names(bin2) <- names(bin1)
set.seed(100)
mass <- runif(10, 10, 800)
beak.size <- runif(10, .2, 5)
tail.length <- runif(10, 2, 10)
wing.length <- runif(10, 15, 60)
range.size <- runif(10, 10000, 100000)
traits <- data.frame(mass, beak.size, tail.length, wing.length, range.size)
rownames(traits) <- names(bin1)
set.seed(100)
tree <- ape::rtree(n = 10, tip.label = names(bin1))
temp.beta(bin1, bin2)
temp.beta(bin1, bin2, traits)
temp.beta(bin1, bin2, tree)
}
}
