% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_temporal.R
\name{temp.beta.vec}
\alias{temp.beta.vec}
\title{Temporal beta diversity calculation for vector}
\usage{
temp.beta.vec(x, nspp, spp, tree, resu)
}
\arguments{
\item{x}{vector with binary distribution of species}

\item{nspp}{species number}

\item{spp}{species name}

\item{tree}{species traits or phylogenetic tree}

\item{resu}{vector to save results}
}
\value{
vector
}
\description{
Temporal beta diversity calculation for vector
}
