% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_size2.R
\name{area.calc.flex}
\alias{area.calc.flex}
\title{Flexible Area Calculation for Raster}
\usage{
area.calc.flex(
  r1,
  r2_raster = NULL,
  r2_vector = NULL,
  threshold = NULL,
  zonal_polys = NULL,
  id_col = NULL,
  add_cols = NULL,
  omit_zero = TRUE,
  unit = "km"
)
}
\arguments{
\item{r1}{The primary SpatRaster with integer categories.}

\item{r2_raster}{An optional SpatRaster for overlay analysis.}

\item{r2_vector}{An optional SpatVector for overlay analysis.}

\item{threshold}{A numeric value required to binarize 'r2_raster' if it's continuous.}

\item{zonal_polys}{An optional SpatVector for zonal analysis.}

\item{id_col}{A string specifying the column in 'zonal_polys'.}

\item{add_cols}{An optional character vector of additional column names from 'zonal_polys'.}

\item{omit_zero}{A logical value. If TRUE (default), results for category = 0 are removed.}

\item{unit}{A string specifying the area unit ("km", "m", or "ha").}
}
\value{
A data frame with the area for each category.
}
\description{
This function calculates the area of integer categories in a primary raster (r1).
It can optionally compute an overlay area with a second layer (r2) and/or
perform calculations within distinct zones defined by a polygon SpatVector.
}
