% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_size2.R
\name{area.calc.flex}
\alias{area.calc.flex}
\title{Calculates Zonal or Total Area for Species Rasters}
\usage{
area.calc.flex(
  r1_multi,
  r2 = NULL,
  polys = NULL,
  threshold,
  category_col = NULL,
  cellSz = NULL
)
}
\arguments{
\item{r1_multi}{A SpatRaster with multiple layers (species), with values of 1 (presence) and 0 (absence).}

\item{r2}{A single-layer SpatRaster. Can be continuous or binary (0/1). Optional.}

\item{polys}{A SpatVector of polygons with categories. Optional.}

\item{threshold}{A numeric threshold to binarize \code{r2}. Only required if \code{r2} is continuous.}

\item{category_col}{The name of the column in \code{polys} with the categories. Required if \code{polys} is used.}

\item{cellSz}{A pre-calculated SpatRaster of cell sizes in km^2. If NULL (default), it will be
calculated automatically. Providing this can increase efficiency.}
}
\value{
A standard R data.frame with the calculated areas in square kilometers.
}
\description{
This function calculates the area of suitable habitat from a binary species raster (\code{r1_multi})
and, optionally, the area of overlap with a second raster (\code{r2}). The second
raster \code{r2} can be continuous (requiring a threshold) or already binary.
The calculation can be performed for the entire raster extent ("total") or be
separated by categories defined in a polygon vector (\code{polys}).
}
