% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_size.R
\name{area.calc}
\alias{area.calc}
\title{Area calculation}
\usage{
area.calc(r1, unit = "km", r2 = NULL, r3 = NULL, filename = "")
}
\arguments{
\item{r1}{A SpatRaster or a list of multiple SpatRasters.}

\item{unit}{Character. Default is "km", but "m" and "ha" are
also available.}

\item{r2}{A SpatRaster with the same resolution and extent as "r1".}

\item{r3}{A SpatRaster with the same resolution and extent as "r1".}

\item{filename}{Character. Save results if a name is provided.}
}
\value{
A vector or a data.frame with area values in the chosen
unit.
}
\description{
Area calculation
}
\examples{
\donttest{
library(terra)
bin1 <- terra::rast(system.file("extdata", "ref.tif",
package = "divraster"))
my.list <- list(bin1 = bin1[[1:2]])
area.calc(my.list)
}
}
