% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecogen.R
\name{ecogen}
\alias{ecogen}
\title{Average trait generalization}
\usage{
ecogen(r, trait, filename = NULL, cores = 1, ...)
}
\arguments{
\item{r}{SpatRaster object with binarized distribution projected to all species in a given climate scenario}

\item{trait}{Numeric vector with trait value for each species}

\item{filename}{Output filename}

\item{cores}{A positive integer indicating if parallel processing should be used (cores > 1)}

\item{...}{Additional arguments to be passed down from a calling function}
}
\value{
SpatRaster object with the average trait chosen
}
\description{
Average trait generalization
}
\examples{
\dontrun{
set.seed(100)
r <- terra::rast(array(sample(c(rep(1, 800), rep(0, 200))), dim = c(10, 10, 10)))
names(r) <- paste0("sp", 1:10)
r
set.seed(100)
mass <- runif(10, 10, 800)
mass
t <- ecogen(r, mass)
t
}
}
