% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecogen.R
\name{ecogen}
\alias{ecogen}
\title{Average trait generalization}
\usage{
ecogen(r, trait)
}
\arguments{
\item{r}{SpatRaster object with binarized distribution projected to all species in a given climate scenario}

\item{trait}{Numeric vector with trait value for each species}
}
\value{
SpatRaster object with the average trait chosen
}
\description{
Average trait generalization
}
\examples{
\dontrun{
set.seed(100)
ref <- terra::rast(array(sample(c(rep(1, 750), rep(0, 250))), dim = c(20, 20, 10)))
set.seed(100)
mass <- runif(10, 10, 800) # grams
mass
t <- ecogen(ref, mass)
t
}
}
