% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecogen.R
\name{ecogen}
\alias{ecogen}
\title{Calculate average trait generalization}
\usage{
ecogen(bin, tab, trait = "one")
}
\arguments{
\item{bin}{Object of class SpatRaster with binarized distribution projected to all species from climate scenario 1}

\item{tab}{Table with species as rows and traits as columns}

\item{trait}{The default parameter of the argument is "one" that calculates ecological generalization to a single trait in the first column of the tab object. Alternatively, the parameter can be set to "all" that calculates ecological generalization to all traits at once}
}
\value{
Object of class SpatRaster with average trait chosen. If trait = "all", it return a list with all traits
}
\description{
Calculate average trait generalization
}
\examples{
\dontrun{
set.seed(100)
ref <- terra::rast(array(sample(c(rep(1, 750), rep(0, 250))), dim = c(20, 20, 10)))
names(ref) <- paste0("sp", 1:10)

set.seed(100)
mass <- runif(10, 10, 800) # grams
beaksize <- runif(10, 5, 30) # millimeters
sp <- paste0("sp", 1:10)
traits <- as.data.frame(cbind(sp, mass, beaksize))
t1 <- ecogen(ref, traits, "one")
t1
t2 <- ecogen(ref, traits, "all")
t2
}
}
