% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{spat.alpha.vec}
\alias{spat.alpha.vec}
\title{Alpha calculation for vector}
\usage{
spat.alpha.vec(x, tree, resu, ...)
}
\arguments{
\item{x}{A numeric vector with presence-absence data (0 or 1)
for a set of species.}

\item{tree}{It can be a data frame with species traits or a
phylogenetic tree.}

\item{resu}{Numeric. A vector to store results.}

\item{...}{Additional arguments to be passed passed down from
a calling function.}
}
\value{
A SpatRaster with alpha result.
}
\description{
Calculates alpha diversity for taxonomic (TD),
functional (FD), and phylogenetic (PD) dimensions.
Adapted from \code{\link[BAT]{alpha}}
}
