% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{spat.alpha.vec}
\alias{spat.alpha.vec}
\title{Alpha calculation for a vector}
\usage{
spat.alpha.vec(x, tree, resu, ...)
}
\arguments{
\item{x}{A SpatRaster with presence-absence data (0 or 1) for a set of species.}

\item{tree}{It can be a phylo object with phylogenetic tree for a set of species or a data.frame object with species traits.}

\item{resu}{numeric vector}

\item{...}{additional arguments}
}
\value{
A numeric vector with alpha result.
}
\description{
Alpha calculation for a vector
}
