% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha2.R
\name{spat.alpha2}
\alias{spat.alpha2}
\title{Alternative method to calculate alpha taxonomic diversity}
\usage{
spat.alpha2(bin, cores = 1, filename = "")
}
\arguments{
\item{bin}{A SpatRaster with presence-absence data (0 or 1) for
a set of species.}

\item{cores}{A positive integer. If cores > 1, a 'parallel'
package cluster with that many cores is created and used.}

\item{filename}{Character. Save results if a name is provided.}
}
\value{
A SpatRaster object with richness.
}
\description{
Alternative method to calculate alpha taxonomic diversity
}
\examples{
\donttest{
library(terra)
bin1 <- terra::rast(system.file("extdata", "ref.tif",
package = "divraster"))
spat.alpha2(bin1)
}
}
