% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_spatial.R
\name{spat.beta.vec}
\alias{spat.beta.vec}
\title{Compute beta diversity on vectors}
\usage{
spat.beta.vec(x, tree, global = FALSE, func = "jac", spp, nspp)
}
\arguments{
\item{x}{numeric vector with multiple species and multiple sites}

\item{tree}{a data.frame with species traits or a phylogenetic tree}

\item{global}{default = FALSE to compare dissimilarity between focal cell and its neighboring cells}

\item{func}{dissimilarity index. default = jaccard}

\item{spp}{species names}

\item{nspp}{numeric. number of species to transform the vector
into a matrix with nspp columns#'}
}
\value{
spatraster object
}
\description{
This function will compute beta diversity on vectors containing
multiple species at multiple sites. Species at sites should be
placed sequentially, so that the vector can be transformed in a
matrix with species at columns and sites at rows
}
