% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_spatial.R
\name{spat.beta.vec}
\alias{spat.beta.vec}
\title{Spatial beta diversity for vector}
\usage{
spat.beta.vec(x, tree, global = FALSE, spp, nspp, ...)
}
\arguments{
\item{x}{A numeric vector with presence-absence data (0 or 1) for a set of species.}

\item{tree}{It can be a data frame with species traits or a phylogenetic tree.}

\item{global}{Logical. Mean of pairwise comparisons between focal cell and its neighbors (default) or mean of all pairwise comparisons.}

\item{spp}{Character. Species names.}

\item{nspp}{Numeric. Number of species.}

\item{...}{Additional arguments to be passed passed down from a calling function.}
}
\value{
A SpatRaster with beta results.
}
\description{
This function will compute beta diversity on vectors containing
multiple species at multiple sites. Species at sites should be
placed sequentially, so that the vector can be transformed in a
matrix with species at columns and sites at rows.
}
