% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_temporal.R
\name{temp.beta}
\alias{temp.beta}
\title{Temporal beta diversity calculation for raster}
\usage{
temp.beta(bin1, bin2, tree, cores = 1, filename = NULL, overwrite = TRUE, ...)
}
\arguments{
\item{bin1}{spatraster}

\item{bin2}{spatraster}

\item{tree}{species traits or phylogenetic tree}

\item{cores}{parallel calculation}

\item{filename}{save file if a name is provided}

\item{overwrite}{replace old file}

\item{...}{additional arguments}
}
\value{
spatraster
}
\description{
Temporal beta diversity calculation for raster
}
\examples{
\dontrun {
set.seed(100)
bin1 <- terra::rast(ncol = 5, nrow = 5, nlyr = 10)
values(bin1) <- round(runif(ncell(bin1) * nlyr(bin1)))
names(bin1) <- paste0("sp", 1:10)
bin2 <- terra::rast(ncol = 5, nrow = 5, nlyr = 10)
values(bin2) <- round(runif(ncell(bin2) * nlyr(bin2)))
names(bin2) <- names(bin1)
set.seed(100)
mass <- runif(10, 10, 800)
beak.size <- runif(10, .2, 5)
tail.length <- runif(10, 2, 10)
wing.length <- runif(10, 15, 60)
range.size <- runif(10, 10000, 100000)
traits <- data.frame(mass, beak.size, tail.length, wing.length, range.size)
rownames(traits) <- names(bin1)
set.seed(100)
tree <- ape::rtree(n = 10, tip.label = paste0("sp", 1:10))
temp.beta(bin1, bin2)
temp.beta(bin1, bin2, traits)
temp.beta(bin1, bin2, tree)
}
}
